fix: Remove broken resources
This drops the resources that no-longer have a struct in Rancher.

This removes the functionality from RKE1 from the terraform provider.
