+++ rancher2/provider.go
@@ -128,10 +128,8 @@ func Provider() terraform.ResourceProvider {
 			"rancher2_cluster_proxy_config_v2":                       resourceRancher2ClusterProxyConfigV2(),
 			"rancher2_cluster_role_template_binding":                 resourceRancher2ClusterRoleTemplateBinding(),
 			"rancher2_cluster_sync":                                  resourceRancher2ClusterSync(),
-			"rancher2_cluster_template":                              resourceRancher2ClusterTemplate(),
 			"rancher2_config_map_v2":                                 resourceRancher2ConfigMapV2(),
 			"rancher2_custom_user_token":                             resourceRancher2CustomUserToken(),
-			"rancher2_etcd_backup":                                   resourceRancher2EtcdBackup(),
 			"rancher2_feature":                                       resourceRancher2Feature(),
 			"rancher2_global_role":                                   resourceRancher2GlobalRole(),
 			"rancher2_global_role_binding":                           resourceRancher2GlobalRoleBinding(),
@@ -139,7 +137,6 @@ func Provider() terraform.ResourceProvider {
 			"rancher2_namespace":                                     resourceRancher2Namespace(),
 			"rancher2_node_driver":                                   resourceRancher2NodeDriver(),
 			"rancher2_node_pool":                                     resourceRancher2NodePool(),
-			"rancher2_node_template":                                 resourceRancher2NodeTemplate(),
 			"rancher2_pod_security_admission_configuration_template": resourceRancher2PodSecurityAdmissionConfigurationTemplate(),
 			"rancher2_project":                                       resourceRancher2Project(),
 			"rancher2_project_role_template_binding":                 resourceRancher2ProjectRoleTemplateBinding(),
@@ -161,15 +158,12 @@ func Provider() terraform.ResourceProvider {
 			"rancher2_cluster_v2":                                    dataSourceRancher2ClusterV2(),
 			"rancher2_cluster_driver":                                dataSourceRancher2ClusterDriver(),
 			"rancher2_cluster_role_template_binding":                 dataSourceRancher2ClusterRoleTemplateBinding(),
-			"rancher2_cluster_template":                              dataSourceRancher2ClusterTemplate(),
 			"rancher2_config_map_v2":                                 dataSourceRancher2ConfigMapV2(),
-			"rancher2_etcd_backup":                                   dataSourceRancher2EtcdBackup(),
 			"rancher2_global_role":                                   dataSourceRancher2GlobalRole(),
 			"rancher2_global_role_binding":                           dataSourceRancher2GlobalRoleBinding(),
 			"rancher2_namespace":                                     dataSourceRancher2Namespace(),
 			"rancher2_node_driver":                                   dataSourceRancher2NodeDriver(),
 			"rancher2_node_pool":                                     dataSourceRancher2NodePool(),
-			"rancher2_node_template":                                 dataSourceRancher2NodeTemplate(),
 			"rancher2_pod_security_admission_configuration_template": dataSourceRancher2PodSecurityAdmissionConfigurationTemplate(),
 			"rancher2_principal":                                     dataSourceRancher2Principal(),
 			"rancher2_project":                                       dataSourceRancher2Project(),
